Move student tabs to single nav destination to prevent flickering from NavHost animation.

+++ app/src/main/java/com/example/teacherapp/ui/screens/student/data/StudentScaffoldViewModel.kt
@@ -1,9 +1,6 @@
 package com.example.teacherapp.ui.screens.student.data
 
 import android.database.sqlite.SQLiteException
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -12,6 +9,7 @@ import com.example.teacherapp.data.models.Resource
 import com.example.teacherapp.data.models.ResourceStatus
 import com.example.teacherapp.data.models.entities.Student
 import com.example.teacherapp.ui.nav.graphs.student.studentIdArg
+import com.example.teacherapp.ui.nav.graphs.student.tab.StudentTab
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.*
@@ -19,16 +17,16 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class StudentContentViewModel @Inject constructor(
+class StudentScaffoldViewModel @Inject constructor(
     private val studentDataSource: StudentDataSource,
-    savedStateHandle: SavedStateHandle,
+    private val savedStateHandle: SavedStateHandle,
 ) : ViewModel() {
 
     private val resourceStatus = MutableStateFlow<ResourceStatus>(ResourceStatus.Loading)
 
     private val studentId = savedStateHandle.getStateFlow(STUDENT_ID_KEY, 0L)
-
-    var isStudentDeleted by mutableStateOf(false)
+    val selectedTab = savedStateHandle.getStateFlow(SELECTED_STUDENT_TAB_KEY, StudentTab.Detail)
+    val isStudentDeleted = savedStateHandle.getStateFlow(IS_STUDENT_DELETED_KEY, false)
 
     @OptIn(ExperimentalCoroutinesApi::class)
     private val student: Flow<Student?> = studentId
@@ -65,17 +63,24 @@ class StudentContentViewModel @Inject constructor(
             initialValue = Resource.Loading,
         )
 
-    fun deleteStudent() {
-        viewModelScope.launch {
-            isStudentDeleted = false
-
-            studentDataSource.deleteStudentById(studentId.value)
-
-            isStudentDeleted = true
+    fun onDeleteStudent() {
+        if (isStudentDeleted.value) {
+            return
         }
+
+        viewModelScope.launch {
+            studentDataSource.deleteStudentById(studentId.value)
+            savedStateHandle[IS_STUDENT_DELETED_KEY] = true
+        }
+    }
+
+    fun onSelectTab(selectedTab: StudentTab) {
+        savedStateHandle[SELECTED_STUDENT_TAB_KEY] = selectedTab
     }
 
     companion object {
         private const val STUDENT_ID_KEY = studentIdArg
+        private const val SELECTED_STUDENT_TAB_KEY = "selected-student-tab"
+        private const val IS_STUDENT_DELETED_KEY = "is-student-deleted"
     }
 }
